backend/services/video_processor.py
"""
Video Processing Service
Handles Kelly's base video analysis and processing for content generation
"""

import os
import cv2
import numpy as np
from typing import Dict, Any, Optional, Tuple
import json
import logging
from pathlib import Path

logger = logging.getLogger(__name__)

class KellyVideoProcessor:
    """
    Processes Kelly's base video for content generation
    """
    
    def __init__(self, base_video_path: str):
        self.base_video_path = base_video_path
        self.video_cap = None
        self.metadata = {}
        self.audio_track = None
        self.sync_points = []
        self.face_landmarks = []
        
    def analyze_video(self) -> Dict[str, Any]:
        """
        Analyze Kelly's base video to extract metadata and sync points
        """
        logger.info(f"Analyzing Kelly's base video: {self.base_video_path}")
        
        if not os.path.exists(self.base_video_path):
            raise FileNotFoundError(f"Kelly's base video not found: {self.base_video_path}")
        
        # Open video file
        self.video_cap = cv2.VideoCapture(self.base_video_path)
        
        if not self.video_cap.isOpened():
            raise ValueError(f"Could not open video file: {self.base_video_path}")
        
        # Extract basic video properties
        self.metadata = {
            "width": int(self.video_cap.get(cv2.CAP_PROP_FRAME_WIDTH)),
            "height": int(self.video_cap.get(cv2.CAP_PROP_FRAME_HEIGHT)),
            "fps": self.video_cap.get(cv2.CAP_PROP_FPS),
            "frame_count": int(self.video_cap.get(cv2.CAP_PROP_FRAME_COUNT)),
            "duration": self.video_cap.get(cv2.CAP_PROP_FRAME_COUNT) / self.video_cap.get(cv2.CAP_PROP_FPS),
            "codec": self._get_video_codec(),
            "file_size": os.path.getsize(self.base_video_path)
        }
        
        # Analyze video content
        self._analyze_face_movement()
        self._extract_audio_track()
        self._identify_sync_points()
        
        # Save metadata
        self._save_metadata()
        
        logger.info(f"Video analysis complete. Duration: {self.metadata['duration']:.2f}s, Resolution: {self.metadata['width']}x{self.metadata['height']}")
        
        return self.metadata
    
    def _get_video_codec(self) -> str:
        """Get video codec information"""
        try:
            # This is a simplified approach - in production, use ffprobe
            return "mp4v"  # Default assumption
        except:
            return "unknown"
    
    def _analyze_face_movement(self):
        """Analyze Kelly's face movement patterns for lip-sync"""
        logger.info("Analyzing face movement patterns...")
        
        # Initialize face detection
        face_cascade = cv2.CascadeClassifier(cv2.data.haarcascades + 'haarcascade_frontalface_default.xml')
        
        frame_count = 0
        face_positions = []
        
        while True:
            ret, frame = self.video_cap.read()
            if not ret:
                break
            
            # Convert to grayscale for face detection
            gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
            
            # Detect faces
            faces = face_cascade.detectMultiScale(gray, 1.1, 4)
            
            if len(faces) > 0:
                # Use the largest face (assuming Kelly is the main subject)
                face = max(faces, key=lambda x: x[2] * x[3])
                face_positions.append({
                    "frame": frame_count,
                    "x": int(face[0]),
                    "y": int(face[1]),
                    "width": int(face[2]),
                    "height": int(face[3])
                })
            
            frame_count += 1
        
        # Reset video to beginning
        self.video_cap.set(cv2.CAP_PROP_POS_FRAMES, 0)
        
        self.face_landmarks = face_positions
        logger.info(f"Detected face in {len(face_positions)} frames")
    
    def _extract_audio_track(self):
        """Extract audio track from video for reference"""
        logger.info("Extracting audio track...")
        
        # For now, we'll use a placeholder approach
        # In production, use ffmpeg to extract audio
        self.audio_track = {
            "duration": self.metadata["duration"],
            "sample_rate": 44100,  # Default assumption
            "channels": 2,  # Stereo
            "format": "wav"
        }
        
        logger.info("Audio track extracted")
    
    def _identify_sync_points(self):
        """Identify key sync points for lip-sync"""
        logger.info("Identifying sync points...")
        
        # Analyze face movement to find key talking points
        if not self.face_landmarks:
            return
        
        # Find frames with significant face movement (talking)
        movement_threshold = 5  # pixels
        sync_points = []
        
        for i in range(1, len(self.face_landmarks)):
            prev_face = self.face_landmarks[i-1]
            curr_face = self.face_landmarks[i]
            
            # Calculate movement
            movement = abs(curr_face["x"] - prev_face["x"]) + abs(curr_face["y"] - prev_face["y"])
            
            if movement > movement_threshold:
                sync_points.append({
                    "frame": curr_face["frame"],
                    "timestamp": curr_face["frame"] / self.metadata["fps"],
                    "movement": movement
                })
        
        self.sync_points = sync_points
        logger.info(f"Identified {len(sync_points)} sync points")
    
    def _save_metadata(self):
        """Save video metadata to file"""
        metadata_file = os.path.join(os.path.dirname(self.base_video_path), "kelly_metadata.json")
        
        metadata = {
            "video_properties": self.metadata,
            "face_landmarks": self.face_landmarks,
            "sync_points": self.sync_points,
            "audio_track": self.audio_track,
            "analysis_timestamp": str(np.datetime64('now'))
        }
        
        with open(metadata_file, 'w') as f:
            json.dump(metadata, f, indent=2)
        
        logger.info(f"Metadata saved to {metadata_file}")
    
    def get_video_segment(self, start_time: float, duration: float) -> str:
        """
        Extract a video segment for content generation
        
        Args:
            start_time: Start time in seconds
            duration: Duration in seconds
            
        Returns:
            Path to extracted video segment
        """
        logger.info(f"Extracting video segment: {start_time}s to {start_time + duration}s")
        
        # Ensure video cap is open
        if not self.video_cap or not self.video_cap.isOpened():
            self.video_cap = cv2.VideoCapture(self.base_video_path)
        
        # Calculate frame numbers
        start_frame = int(start_time * self.metadata["fps"])
        end_frame = int((start_time + duration) * self.metadata["fps"])
        
        # Create output path
        output_dir = os.path.join(os.path.dirname(self.base_video_path), "segments")
        os.makedirs(output_dir, exist_ok=True)
        
        output_path = os.path.join(output_dir, f"kelly_segment_{start_time:.2f}_{duration:.2f}.mp4")
        
        # Extract segment using OpenCV
        self.video_cap.set(cv2.CAP_PROP_POS_FRAMES, start_frame)
        
        # Get video writer
        fourcc = cv2.VideoWriter_fourcc(*'mp4v')
        out = cv2.VideoWriter(output_path, fourcc, self.metadata["fps"], 
                            (self.metadata["width"], self.metadata["height"]))
        
        # Write frames
        for frame_num in range(start_frame, min(end_frame, self.metadata["frame_count"])):
            ret, frame = self.video_cap.read()
            if not ret:
                break
            out.write(frame)
        
        out.release()
        
        logger.info(f"Video segment saved to {output_path}")
        return output_path
    
    def create_talking_head_video(self, audio_path: str, output_path: str) -> str:
        """
        Create a new talking head video with Kelly using provided audio
        
        Args:
            audio_path: Path to generated audio file
            output_path: Path for output video
            
        Returns:
            Path to created video
        """
        logger.info(f"Creating talking head video with audio: {audio_path}")
        
        # This is a simplified implementation
        # In production, use advanced lip-sync technology like Wav2Lip
        
        # For now, we'll create a basic video by:
        # 1. Taking a segment from Kelly's base video
        # 2. Replacing the audio track
        # 3. Saving as new video
        
        # Get audio duration
        audio_duration = self._get_audio_duration(audio_path)
        
        # Extract video segment of same duration
        video_segment = self.get_video_segment(0, audio_duration)
        
        # Combine video and audio (simplified approach)
        # In production, use ffmpeg for proper audio-video sync
        final_video = self._combine_audio_video(video_segment, audio_path, output_path)
        
        logger.info(f"Talking head video created: {final_video}")
        return final_video
    
    def _get_audio_duration(self, audio_path: str) -> float:
        """Get duration of audio file"""
        # Simplified approach - in production, use librosa or similar
        return 5.0  # Placeholder
    
    def _combine_audio_video(self, video_path: str, audio_path: str, output_path: str) -> str:
        """Combine video and audio into final talking head video"""
        # Simplified approach - in production, use ffmpeg
        # For now, just copy the video segment
        import shutil
        shutil.copy2(video_path, output_path)
        return output_path
    
    def cleanup(self):
        """Clean up resources"""
        if self.video_cap:
            self.video_cap.release()

def analyze_kelly_video(video_path: str) -> Dict[str, Any]:
    """
    Analyze Kelly's base video and return metadata
    
    Args:
        video_path: Path to Kelly's base video file
        
    Returns:
        Video metadata and analysis results
    """
    processor = KellyVideoProcessor(video_path)
    
    try:
        metadata = processor.analyze_video()
        return {
            "success": True,
            "metadata": metadata,
            "face_landmarks_count": len(processor.face_landmarks),
            "sync_points_count": len(processor.sync_points)
        }
    except Exception as e:
        logger.error(f"Error analyzing Kelly's video: {e}")
        return {
            "success": False,
            "error": str(e)
        }
    finally:
        processor.cleanup()

if __name__ == "__main__":
    # Test the video processor
    video_path = "../kelly2-welcome-pitch.mp4"
    
    if os.path.exists(video_path):
        result = analyze_kelly_video(video_path)
        print("Kelly Video Analysis Result:")
        print(json.dumps(result, indent=2))
    else:
        print(f"Video file not found: {video_path}")
