backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
openai==1.3.7
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
dj-database-url==2.1.0